+++ torch_frame/gbdt/tuned_lightgbm.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import Any
+from typing import Any, Optional
 
 import numpy as np
 import pandas as pd
@@ -221,3 +221,27 @@ class LightGBM(GBDT):
         import lightgbm
 
         self.model = lightgbm.Booster(model_file=path)
+
+    def _feature_importance(self, importance_type: str = 'gain',
+                            iteration: Optional[int] = None) -> list:
+        r"""Get feature importances.
+
+        Args:
+        importance_type (str): How the importance is calculated.
+            If "split", result contains numbers of times the feature
+            is used in a model. If "gain", result contains total gains
+            of splits which use the feature.
+        iteration (int, optional): Limit number of `iterations` in the feature
+            importance calculation. If None, if the best `iteration` exists,
+            it is used; otherwise, all trees are used. If <= 0, all trees
+            are used (no limits).
+
+        Returns:
+            list: Array with feature importances.
+        """
+        assert importance_type in [
+            'split', 'gain'
+        ], f'Expect split or gain, got {importance_type}.'
+        scores = self.model.feature_importance(importance_type=importance_type,
+                                               iteration=iteration)
+        return scores.tolist()